Don't hardcode the term start months
This patch gets rid of the hardcoded term start months (april and october) and makes them configurable within two admin settings.

In addition to that change, two functions are added to the data_controller class which make the term start months gettable for callers from outside of the plugin.
To make clear from when on these functions are available for other plugins which require this plugin, the plugin version is bumped.