Allow admin to define the sort order of the field values.
Before this patch, the sort ortder of the field values (as they appear within a course's settings and, if configured, in the course list filter on block_myoverview) was hardcoded to shw older terms first and add the term-independent term at the beginning of the list.
The more terms you have and the longer the list grows, the more this sorting becomes counterintuitive.

This patch adds an admin settings which lets the admin control how the list of terms is presented within the course settings and (if the field is used as a course filter) within the Dashboard course overview block.